Refactor: introduce `Version.Update` as a parameter object
The `currentVersion` & `nextVersion` parameters are paired together
relatively often, and because they are of the same type (`Version`)
they are susceptible to being accidentally swapped over without any
compiler error being raised. Encapsulating them into a single
concept (that of a `Version.Update`) makes code a little clearer,
and reduces the number of individual parameters being passed on
method calls.