Do not cache the output of `gitAskPass`
Prior to this change, the `gitAskPass` program was called once on
start-up and its output was cached and used for API calls to the forge
while Git itself does not cache its output but calls it anytime a password
is needed. This means if the output of `gitAskPass` changes during a Scala
Steward run, the new password is only used for Git operations but not for
forge API calls.

With this change we now do the same as Git and call `gitAskPass`
everytime the password is needed. This should make it easier to support
GitHub Apps proper which require different access tokens during a run.
See also https://github.com/scala-steward-org/scala-steward/issues/2973#issuecomment-1866224921.