Remove repo from function for modifying requests
This removes the unnecessary `Repo` parameter from the function for
modifying requests to add authentication headers. This parameter was
added in #373 but wasn't ever used as can be observed in the diff of
`ForgeSelection`. My motivation for adding `Repo` back then was to
eventually use different credentials for different repositories
for proper GitHub App support. But #3250 demonstrates that we can
achieve this without this extra parameter.