Refactor pull request head: replace conditionals with polymorphism
This is another example of replacing conditionals with polymorphism on
`ForgeType`. This was a theme in commit b42866be, merged in this PR back
in August 2023:

https://github.com/scala-steward-org/scala-steward/pull/3145

A benefit of this pattern is as new types of forge are implemented, the
responsibilities for supporting that new `ForgeType` are all in one
place, rather than spread across several methods elsewhere in the
codebase.

To confess, I think there's probably a further refactor that could be
done, to introduce a better parameter object denoting a 'Pull Request
Head' (that conceivably could better model the way that some
`ForgeType`s don't even support forking) - but that can wait for
another day, fixing #3300 is more urgent!