Fix the merging order updates.pin
The merging of the default Scala Steward configuration for pinned artifacts was reversed.

In contrast to the comment and intend:
```Scala
//  Strategy: union with repo preference in terms of revision
```

This pull-request fixes the merging order and adapts the test.
In addition it adds tests that show how `updates.pin` can  be used to control Scala LTS release updates.