Combine `RepoConfig` with the one taken the higher importance on the right side.
This is done in the `RepoConfigLoader` with the default config and the supplied additional configs.
And now here in the same way.

The `RepoConfig.combine`  function treats the 2nd argument as the one overwriting the 1st one in cas emerging doesn't make sense. This applies at the moment only for "pin".