Allow configuring exit code behaviour
The new `--exit-code-success-if-any-repo-succeeds` CLI flag alters the default
behaviour of the exit-code of the Scala Steward process - the exit code will be
success (`0`) if the processing of _any_ repo has succeeded, rather than
requiring that _all_ repos are successfully processed. This is useful when running
Scala Steward with the Scala Steward GitHub Action - it means that the filesystem
_will_ be persisted, along with the record of what PRs have been raised, which is
what you _want_ to happen, even if some repos have failed.

As Alejandro mentioned:

https://github.com/scala-steward-org/scala-steward/pull/3369#issuecomment-2151485229

...this behaviour may not always be desired, so we're disabling it by default.