Make all values in config classes optional (#3515)
This changes `RepoConfig` and the other `*Config` classes such that all
values are optional and there are no default parameter values except
`None`s. This allows us to get rid of circe-generic-extras'
`Configuration.default.withDefaults` and `deriveConfiguredCodec` since
missing `Option` values do not cause an error during decoding. So `{}`
can still be decoded into a `RepoConfig`, and there are a lot of tests
in `RepoConfigAlgTest` that ensure that sparse configurations with a
lot of missing values can still be decoded.