Add dependency-cooldowns / minimum-artifact-age
This PR adds a new `updates.cooldown` config option to allow configuring Scala Steward to defer suggesting an update to a new version of a dependency, until that version has survived a 'cooldown' period (eg 7 days).

This intended to be useful to users who want to reduce the risk of updating to a malicious new version of a dependency before it's been examined by their supply-chain security vendor, addressing:

* https://github.com/scala-steward-org/scala-steward/issues/3757

Note that applying a cooldown does not necessarily mean that you will get _fewer_ updates - just that any update offered will be at least as old as specified.

Artifacts that are too recent will get rejected with a `TooRecentForCooldown` rejection reason.

## Internal changes

* `VersionsCache` now persists `VersionWithFirstSeen` rather than `Version`, allowing us to record when a new version was first seen by this Scala Steward instance (the only simple & reliable way to find out how old a Maven artifact is!). We've added `VersionsCacheTest` to ensure we can still decode both old and new versions of the cache data.

## Config

```
updates.cooldown = {
  minimumAge: "7 days"
}
```

Co-authored-by: Roberto Tyley <roberto.tyley@guardian.co.uk>