Redux: Dependency Cooldowns, post #3787
Add deserialisation test for VersionsCache Value

Add json fixture for new Value type

Add VersionWithFirstSeen

Add merging of new and old first-seen version data

Add deserialisation test for new format

It looks like we probably want to update UpdatesConfig.scala to add a
new method `isTooNew: Update.ForArtifactId => FilterResult` which rules
out proposed updates that use too new a version (according to the user’s
config).

To do that, we reckon we’ll need Update.ForArtifactId to list versions
with their firstSeen information, i.e. update the contained type from
Version to VersionWithFirstSeen. This commit begins making that
change (and so will not compile).

Replace withNewerVersions with supersedes

This commit removes withNewerVersions and replaces the one usage of it
with a new function supersedes, which checks whether an update matches
another while having a newer nextVersion.

withNewerVersions was a little awkward to update to work with us having
changed some Versions to VersionWithFirstSeens, because it accepts a
list of Versions (rather than just producing one).

It was originally added in response to [this PR
comment](https://github.com/scala-steward-org/scala-steward/pull/1667#discussion_r506648775),
but we reckon replacing it with a slightly different check of the
group and artifact IDs makes for clearer, more explicit code. We
think (though it’s somewhat hard to tell) that this doesn’t remove any
existing functionality: part of the argument for adding
withNewerVersions in the first place was to support Update.Group, but it
appears to only be called for Update.Singles.

Add CooldownConfig

To facilitate using dependency cooldowns, this commit adds a class
CooldownConfig to represent a user’s desired cooldown settings: a list
of ages (specified as finite durations, using the same string syntax as
supported for pullRequests.frequency), each with a list of
UpdatePatterns for the age to apply to.

Co-authored-by: Roberto Tyley <roberto.tyley@guardian.co.uk>