Begin propagating firstSeen
It looks like we probably want to update UpdatesConfig.scala to add a
new method `isTooNew: Update.ForArtifactId => FilterResult` which rules
out proposed updates that use too new a version (according to the user’s
config).

To do that, we reckon we’ll need Update.ForArtifactId to list versions
with their firstSeen information, i.e. update the contained type from
Version to VersionWithFirstSeen. This commit begins making that
change (and so will not compile).